requirements.txt
pyautogui
pillow
pystray